Add retain success-builds feature
This change adds a feature of protecting success build result from
being deleted. For this feature to be enabled, a successs-build
directory must exist inside the "log-path-dir" directory. Then the
user must create a symlink inside of success-build directory poiting
to the build dir which wants to protect, or not to be deleted.

Example
> ln -sf {{ zuul_logserver_root }}/{{ zuul_log_path }} /var/www/logs/success-builds/{{ zuul.job }}

Change-Id: Ic0e66af1e60349d54d6d73bfbbf1acaf9b803296